add new class for ship and test
