+++ lib/ship.rb
@@ -1,0 +1,7 @@
+require 'pry'
+
+class Ship
+
+
+
+end
